requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.20.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
attrs==23.2.0
babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
braceexpand==0.1.7
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
configargparse==1.7
cycler==0.11.0
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
flask==3.0.2
fonttools==4.38.0
fqdn==1.5.1
fvcore==0.1.5.post20210915
gitdb==4.0.11
gitpython==3.1.42
google-auth==2.28.1
google-auth-oauthlib==0.4.6
grpcio==1.62.0
gym-notices==0.0.8
idna==3.6
ifcfg==0.21
imageio==2.34.0
imageio-ffmpeg==0.4.9
imgviz==1.7.5
importlib-metadata==7.0.1
importlib-resources==6.1.1
iopath==0.1.9
ipykernel==6.29.2
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
joblib==1.3.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==7.4.9
jupyter-core==5.7.1
jupyter-events==0.9.0
jupyter-packaging==0.12.3
jupyter-server==2.12.5
jupyter-server-fileid==0.9.1
jupyter-server-terminals==0.5.2
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.7
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.3
jupyterlab-widgets==3.0.10
kiwisolver==1.4.4
llvmlite==0.41.1
lmdb==1.0.0
markdown==3.5.2
markupsafe==2.1.5
mathutils==2.81.2
matplotlib==3.7.5
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.7
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nbstripout==0.7.1
nest-asyncio==1.6.0
networkx==3.1
notebook==6.5.6
notebook-shim==0.2.4
numba==0.58.1
numpy==1.22.4
numpy-quaternion==2023.0.2
multipledispatch
oauthlib==3.2.2
objectio==0.2.29
open3d==0.18.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
overrides==7.7.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
plotly==5.19.0
portalocker==1.4.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11-stubgen==2.5.1
pycparser==2.21
pyglet==1.5.28
pygments==2.17.2
pyparsing==3.0.9
pyquaternion==0.9.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pywavelets==1.4.1
pyyaml==6.0
pyzmq==24.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rospkg==1.5.1
rpds-py==0.18.0
rsa==4.9
scikit-image==0.17.2
scikit-learn==1.3.2
scipy==1.7.3
send2trash==1.8.2
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.3.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4
tqdm==4.65.0
traitlets==5.14.1
trimesh==4.3.2
triton==2.0.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webdataset==0.1.40
webencodings==0.5.1
websocket-client==1.7.0
werkzeug==3.0.1
widgetsnbextension==4.0.10
y-py==0.6.2
yacs==0.1.8
ypy-websocket==0.8.4
zipp==3.17.0